Doom + PVM refactor (#526)
Doom demo: https://www.youtube.com/watch?v=ltWMjg2EHS8

Due to performance limitations, the blob is executed directly in PVM instead of through workitems.

To replicate Doom:
```bash
cd pvm && make doom
```

This refactor splits PVM into several modular components:
1. `pvm`: core opcode logic
2. `memory`: memory model
3. `instructions`: opcode number definitions
4. `jam_pvm`: helper functions (e.g. refine, accumulate, on_transfer)

Future improvements:
1. Optimize the memory model
2. Reduce performance bottlenecks caused by large switch statements

+++ pvm/host_testvector_test.go
@@ -1392,7 +1392,7 @@ func ConvertToRefineM_mapForTest(refineM_map map[uint32]*RefineM) map[uint32]*Re
 		}
 
 		for page_addr, page := range v.U.Pages {
-			RAMForTest.Pages[page_addr] = &PageForTest{
+			RAMForTest.Pages[uint32(page_addr)] = &PageForTest{
 				Value:  page.Value,
 				Access: page.Access,
 			}
